rename `Routing` -> `apps-routing`, update aliases
