+++ src/libs/apps-routing/src/assets.ts
