+++ tsconfig.base.json
@@ -9,8 +9,8 @@
       "@polkadot/apps": ["polkadot/packages/apps/src"],
       "@polkadot/apps-config": ["src/libs/apps-config/src"],
       "@polkadot/apps-config/*": ["src/libs/apps-config/src/*"],
-      "@polkadot/apps-routing": ["polkadot/packages/apps-routing/src"],
-      "@polkadot/apps-routing/*": ["polkadot/packages/apps-routing/src/*"],
+      "@polkadot/apps-routing": ["src/libs/apps-routing/src"],
+      "@polkadot/apps-routing/*": ["src/libs/apps-routing/src/*"],
       "@polkadot/app-accounts": ["polkadot/packages/page-accounts/src"],
       "@polkadot/app-accounts/*": ["polkadot/packages/page-accounts/src/*"],
       "@polkadot/app-addresses": ["polkadot/packages/page-addresses/src"],
