+++ webpack.base.cjs
@@ -30,7 +30,7 @@ function createWebpack(context, mode = "production") {
 	const pkgJson = require(path.join(context, "package.json"));
 	const alias = findPackages().reduce((alias, { dir, name }) => {
 		alias["@"] = path.resolve(context, "./src");
-		if (name.includes("apps-config") || name.includes("app-settings")) {
+		if (name.includes("apps-config") || name.includes("app-settings") || name.includes("apps-routing")) {
 			alias[name] = path.resolve(context, `./src/libs/${dir}/src`);
 		} else {
 			alias[name] = path.resolve(context, `./polkadot/packages/${dir}/src`);
